Add `withCacheInstance` parameter and change called argument
It is used to tell if callback is to be called with cache instance object as one of the args data or not.
cache instance, if e.g. console.log'ed, will shadow what are the rest of the arguments, thus it's optional.

As for the called argument - it was changed to object, to tell what event triggered it.